// src/app/models/registrocondominos.model.ts
export interface Registro{
    nombre: string;
    domicilio: string;
    telefono: string;
    tipo: string;
    actdesmsn: string;
    actdesmark: string;
    actdesbal: string;
    actdesenc: string;
    actdescomu: string;
    actdespay: string;
    actdesacc: string;
    actdesos: string;
    actdesreser: string;
    accede: string;
}


export interface Tipo{ 
    tipo: string;
}